pyproject.toml
[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[project]
name = "streamlit-example-app"
version = "0.1.0"
description = "AWS CDK infrastructure for deploying Streamlit apps with Cognito auth"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.180.0",
    "constructs>=10.0.0,<11.0.0",
    "gds-idea-cdk-constructs @ git+https://github.com/co-cddo/gds-idea-cdk-constructs@v.0.1.2",
    "pydantic-settings>=2.0.0",
    "tomlkit>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
]

[project.scripts]
configure = "template.configure:main"
smoke_test = "template.smoke_test:main"

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=6.2.5",
]

[tool.uv.build-backend]
module-name = "template"
module-root = ""

[tool.webapp]
# Web application configuration
app_name = "fastapi-test"
framework = "fastapi"
